Moved treeRotate files into tree/rotate. This was done in preparation for tree/insert and tree/delete files.
